orderer and an org on a single node dns listener fix

+++ host-create.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 source lib/util/util.sh
+source lib.sh
 
 setDocker_LocalRegistryEnv
 
@@ -17,10 +18,11 @@ info "Create Network with vm names prefix: $VM_NAME_PREFIX"
 
 : ${DOMAIN:=example.com}
 
-orgs=${@:-org1}
+declare -A -g ORGS_MAP; parseOrganizationsForDockerMachine ${@:-org1}
+orgs=(`getCurrentOrganizationsList`)
+first_org=${orgs[0]:-org1}
 
 # Create orderer host machine
-
 ordererMachineName="orderer.$DOMAIN"
 
 info "Creating $ordererMachineName, Options: $DOCKER_MACHINE_FLAGS"
@@ -32,9 +34,9 @@ docker-machine create ${DOCKER_MACHINE_FLAGS} ${ordererMachineName}
 
 for org in ${orgs}
 do
-    orgMachineName="${org}.$DOMAIN"
-    info "Creating member organization machine: $orgMachineName with flags: $DOCKER_MACHINE_FLAGS"
-    docker-machine rm ${orgMachineName} --force
+    orgMachineName=`getDockerMachineName $org`
+    info "Creating member organization $org on machine: $orgMachineName with flags: $DOCKER_MACHINE_FLAGS"
+    [ -z "${ORGS_MAP[$org]}" ] && docker-machine rm ${orgMachineName} --force
     docker-machine create ${DOCKER_MACHINE_FLAGS} ${orgMachineName}
 done
